Kafka listener bean post processor decomposition: relocate property resolving methods to a suitable class 'KafkaEndpointPropertyResolver'.
